pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stonkmap"
version = "0.1.0"
description = "Self-hosted portfolio heatmaps for stocks inside index portfolios."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "uvicorn>=0.35.0,<1.0.0",
  "yfinance>=0.2.65,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
]

[project.scripts]
stonkmap-api = "stonkmap.main:run"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
